CI-49 Double Import of the EditPage routes
- Moves the routing out of the edit-routing module and individual page modules
into the single set of routes defined for the lazily-loaded Edit Page Module.

The App module itself only contains formatting changes.
