+++ projects/ranger/src/app/app.module.ts
@@ -5,8 +5,14 @@ import {RouteReuseStrategy} from '@angular/router';
 import {SplashScreen} from '@ionic-native/splash-screen/ngx';
 import {StatusBar} from '@ionic-native/status-bar/ngx';
 
-import {IonicModule, IonicRouteStrategy} from '@ionic/angular';
-import {AuthModule, HeadingModule} from 'cr-lib';
+import {
+  IonicModule,
+  IonicRouteStrategy
+} from '@ionic/angular';
+import {
+  AuthModule,
+  HeadingModule
+} from 'cr-lib';
 import {AppRoutingModule} from './app-routing.module';
 
 import {AppComponent} from './app.component';
