+++ projects/ranger/src/app/edit/draft/draft.module.ts
@@ -1,10 +1,6 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
-import {
-  RouterModule,
-  Routes
-} from '@angular/router';
 
 import {IonicModule} from '@ionic/angular';
 import {
@@ -14,13 +10,6 @@ import {
 
 import {DraftPage} from './draft.page';
 
-const routes: Routes = [
-  {
-    path: '',
-    component: DraftPage
-  }
-];
-
 @NgModule({
   imports: [
     CommonModule,
@@ -28,7 +17,6 @@ const routes: Routes = [
     FormsModule,
     IonicModule,
     LocTypeChipModule,
-    RouterModule.forChild(routes)
   ],
   declarations: [DraftPage]
 })
