+++ projects/ranger/src/app/edit/images/images.module.ts
@@ -1,28 +1,16 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
-import {
-  RouterModule,
-  Routes
-} from '@angular/router';
 
 import {IonicModule} from '@ionic/angular';
 
 import {ImagesPage} from './images.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ImagesPage
-  }
-];
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    RouterModule.forChild(routes)
   ],
   declarations: [ImagesPage]
 })
