+++ projects/ranger/src/app/edit/puzzle/puzzle.module.ts
@@ -1,28 +1,16 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
-import {
-  RouterModule,
-  Routes
-} from '@angular/router';
 
 import {IonicModule} from '@ionic/angular';
 
 import {PuzzlePage} from './puzzle.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: PuzzlePage
-  }
-];
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    RouterModule.forChild(routes)
   ],
   declarations: [PuzzlePage]
 })
